Fix wrong tokens for 3+ layer deep table indexes
When you index a table that's not a built-in one with one that is a built-in twice will make it token as a built-in. I.e.
a.Instance.new();
"a" and "Instance" would token as an iden which is good but "new" would token as a builtin because "Instance" is in the built-in table.